Now simplified, with comments for what needs to be modified
